% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAEPprimer-package.R
\docType{package}
\name{NAEPprimer-package}
\alias{NAEPprimer-package}
\title{The NAEP Primer package contains a random mini-sample of real data from the 2005 grade 8 mathematics
assessment that have been approved for public use. This includes two data files, as well
as a system control file that corresponds to each data file to read in and format for further processing.}
\description{
Most of the data is in \code{\link{M36NC2PM.dat}} (including student, teacher, and most school variables) while 
\code{\link{M36NT2PM.dat}} with two school variables in it for demonstration.
\code{\link{M36NC2PM.fr2}} and \code{\link{M36NT2PM.fr2}}
are format files used to to read in and format for further processing.
}

